feat: cookie logic& adjust advantage of ssr
